(ab)using a custom field to trigger a fix for some websites
By removing the submit handler on the login form auto-fill can be fixed for a few websites (eg. netflix.com)
This is a creative way of triggering this for user selectable websites.
As a single event handler can not be remove all forms are copied, deleted and re-inserted in the original place. This removes all event handler on them.